text2image-app/requirements.txt
gradio>=6.0.2
torch>=2.0.0
diffusers>=0.17.0
transformers>=4.30.0
accelerate>=0.20.0
xformers; platform_system!="Windows"  # optionnel, active si disponible
Pillow
requests
uvicorn[standard]
python-multipart
